requirements.txt
-i https://pypi.org/simple
alembic==1.4.2
bcrypt==3.1.7
blinker==1.4
boto3==1.14.2
botocore==1.17.2
cffi==1.14.0
click==7.1.2
dnspython==1.16.0
docutils==0.15.2
email-validator==1.1.1
flask-bcrypt==0.7.1
flask-email==1.4.4
flask-login==0.5.0
flask-mail==0.9.1
flask-migrate==2.5.3
flask-script==2.0.6
flask-sqlalchemy==2.4.3
flask-wtf==0.14.3
flask==1.1.2
git+git://github.com/e-dard/flask-s3.git@b8c72b40eb38a05135eec36a90f1ee0c96248f72#egg=flask-s3
gunicorn==20.0.4
idna==2.9
install==1.3.1
itsdangerous==1.1.0
jinja2==2.11.2
jmespath==0.10.0
mako==1.1.3
markupsafe==1.1.1
pillow==7.1.2
psycopg2-binary==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
s3transfer==0.3.3
six==1.15.0
sqlalchemy==1.3.17
urllib3==1.25.9 ; python_version != '3.4'
werkzeug==1.0.1
wtforms==2.3.1
